+++ link-grammar/api-types.h
@@ -33,14 +33,12 @@ typedef struct Connector_struct Connector;
 typedef struct Cost_Model_s Cost_Model;
 typedef struct Domain_s Domain;
 typedef struct DTreeLeaf_s DTreeLeaf;
-typedef struct Exp_list_s Exp_list;
 typedef struct Image_node_struct Image_node;
 typedef struct Linkage_info_struct Linkage_info;
 typedef struct Parse_info_struct *Parse_info;
 typedef struct Postprocessor_s Postprocessor;
 typedef struct PP_data_s PP_data;
 typedef struct PP_info_s PP_info;
-typedef struct Regex_node_s Regex_node;
 typedef struct Resources_s * Resources;
 
 /* Some of the more obscure typedefs */
