+++ link-grammar/dict-common/build-disjuncts.h
