+++ link-grammar/dict-file/word-file.c
@@ -11,7 +11,7 @@
 /***************************************************************************/
 
 #include "error.h"
-#include "dict-common.h"
+#include "dict-common/dict-common.h"
 #include "string-set.h"
 #include "structures.h"
 #include "read-dict.h"
