+++ link-grammar/dict-sql/read-sql.c
@@ -20,11 +20,11 @@
 #include <sqlite3.h>
 
 #include "api-structures.h"
-#include "dict-api.h"
-#include "dict-common.h"
-#include "dict-structures.h"
+#include "dict-common/dict-api.h"
+#include "dict-common/dict-common.h"
+#include "dict-common/dict-structures.h"
 #include "externs.h"
-#include "spellcheck.h"
+#include "tokenize/spellcheck.h"
 #include "string-set.h"
 #include "structures.h"
 #include "utilities.h"
