+++ link-grammar/sat-solver/sat-encoder.cpp
@@ -31,8 +31,8 @@ extern "C" {
 
 extern "C" {
 #include "analyze-linkage.h"      // for compute_link_names()
-#include "build-disjuncts.h"      // for build_disjuncts_for_exp()
-#include "dict-api.h"             // for print_expression()
+#include "dict-common/build-disjuncts.h" // for build_disjuncts_for_exp()
+#include "dict-common/dict-api.h" // for print_expression()
 #include "linkage.h"
 #include "post-process/post-process.h"
 #include "post-process/pp-structures.h"
